requirements.txt
fastapi
celery
python-dotenv
redis